requirements.txt
pydantic>=2.0,<3.0
httpx>=0.27,<1.0
neo4j>=5.0,<6.0
tenacity>=8.0,<10.0
python-dotenv>=1.0,<2.0
